Modify radarchart

+++ matte/matteapp/static/matteapp/js/radarChart.js
@@ -99,7 +99,7 @@ function RadarChart(id, data, options) {
 	   .attr("dy", "0.4em")
 	   .style("font-size", "10px")
 	   .attr("fill", "#737373")
-	   .text(function(d,i) { return Format(int(maxValue * d/cfg.levels)); });
+	   .text(function(d,i) { return Format(maxValue * d/cfg.levels); });
 
 	/////////////////////////////////////////////////////////
 	//////////////////// Draw the axes //////////////////////
@@ -223,7 +223,7 @@ function RadarChart(id, data, options) {
 			tooltip
 				.attr('x', newX)
 				.attr('y', newY)
-				.text(Format(int(d.value)))
+				.text(Format(d.value))
 				.transition().duration(200)
 				.style('opacity', 1);
 		})
